Say what number a result answers in one place, and let the readers read it
Two answers to one question. `Question.numberAnsweredBy` read the shape of a
result and answered the number a union carries at exactly one case;
`NumericAnswers.typeOf` read the operation, answered the case an arithmetic
names where one is declared, and otherwise handed back the result type as it
stood. They agree wherever an arithmetic declares its case and part where a
union carries a number and none is declared -- `String.toInt`, where the
first says `Int` and the second says `Int | NotANumber`. Neither of them
lives anywhere whose job is that classification: one is inside what asks the
questions and the other inside what a term's carrier is taken from, so a
third reader has to borrow from a consumer, and the resolver borrowed from
the questions.

So the classification moves into `NumericAnswers`, which already declared
itself the one entry, and the range, the holds and the readings all read it.

What an arithmetic names its case at no longer decides what number an
operation answers. That is a proposition about one representation: an
operation answering `Int | Decimal | SomeError` carries two numbers whatever
its arithmetic names, and reading the number off the arithmetic would make
one representation's statement the answer for all of them. The two are held
to each other where the arithmetic is declared, so a case naming a number the
shape does not carry is refused rather than believed.

`isANumber` goes with it: left behind, the classification would have read the
questions to answer what a number is.
